intent_classifier/requirements.txt
fastapi
uvicorn
pyyaml
groq
pydantic
